requirements.txt
amqp
arrow
asgiref
beautifulsoup4
billiard
bs4
celery==4.4.7
certifi
chardet
click
click-didyoumean
click-plugins
click-repl
Django
django-celery-beat
django-celery-results
django-redis
django-timezone-field
djangorestframework
dnspython
eventlet
flower
greenlet
humanize
ics
idna
kombu
prometheus-client
prompt-toolkit
psycopg2-binary
python-crontab
python-dateutil
pytz
redis
requests
six
soupsieve
sqlparse
TatSu
tornado
tqdm
urllib3
vine
wcwidth
